requirements.txt
asgiref==3.2.3
astroid==2.4.0
certifi==2019.11.28
colorama==0.4.3
Django==3.0.1
django-filter==2.4.0
djangorestframework==3.12.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
pylint==2.5.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
wincertstore==0.2
wrapt==1.12.1
